Create a CLVM storage adaptor
